Update scan request validation to handle disabled plugins
This updates the scan request validation in the repo lambda so that if a
known, but disabled, plugin is explicitly enabled it returns a
meaningful validation error and if a known, but disabled, plugin is
explicitly disabled it is ignored.

The CI shell script explicitly disables the Aqua plugin by default.
However, if the Aqua plugin were disabled by the site deployment this
would result in a validation error. This change fixes that so it is
instead ignored.