Add plugin for gathering secrets from GHAS Secret Scanning
This is a plugin to retrieve and report secrets identified by the GitHub
Advanced Security Secret Scanning. As such, it only works against
repositories that are in GitHub and have Secret Scanning enabled. To
reduce friction, this plugin is only enabled when the feature is enabled
in the Artemis environment, similar to how Veracode, Snyk, and Aqua are
managed.

If the feature is enabled and the repository is in GitHub but the Secret
Scanning functionality is not enabled on the repository the plugin runs
but returns no results. Similarly, if the feature is enabled but the
repository is not in GitHub the plugin runs but returns no results.

In order to function successfully, the GitHub app or PAT used by Artemis
to access GitHub must have sufficient permissions to read the secret
scanning alerts:
https://docs.github.com/en/rest/overview/permissions-required-for-github-apps#secret-scanning-alerts