Stopgap measure so that new secret types aren't filtered out
The ghas_secrets plugin introduced a large number of new secret
types. The list of types will expand as GitHub adds additional
detections:
https://docs.github.com/en/code-security/secret-scanning/secret-scanning-patterns

Secret filtering via the API assumes the type is one of a small
number of static types. This was causing the new secrets from the
ghas_secrets plugin to get filtered out of the results because they
didn't match the expected types. As a workaround, this boolean is
used to bypass the type filtering when no types are specified.

A better solution would be to store a list of secret types in the
database and update it based on the results of plugins so that when
new detections are added the list is updated automatically. This
solution will be implemented at a later time.