Create "Configuration" plugin type and Github Repo Health check (#27)
* Creates a "Configuration" plugin type for auditing configurations of
things.

The "configuration" plugin type represents a pass/fail checklist of rules.
This is particularly useful for checking configurations of things, such
as if an S3 bucket is configured properly, if configurations exist in a
code repository, or if a code repository meets a defined baseline.

* Creates a Github Repo Health plugin for comparing a Github
* repository's configuration against a secure baseline

This is a plugin to compare a Github branch and repository against a
defined, secure baseline. It includes rules for:
    * Ensuring commit signing is required
    * Ensuring that branch protection rules apply to admins
    * Ensuring that branch protection rules that require pull requests to
      meet certain criteria
    * Ensuring that pull requests pass expected status checks
    * Ensuring that actions are configured a certain way
    * Ensuring that Github Advanced Security Code Scanning is enabled
    * Ensuring that a given file (or files) exists (or does not exist) on
      the repo
    * Ensuring that Github Advanced Security Secret Scanning is enabled
    * Ensuring that Dependabot vulnerability alerts are enabled

These rules are configured via JSON objects that are stored as
PluginConfigs (or via the default ruleset stored in the `main.py` for the
plugin). For information on how to configure these JSON objects, view the
readme for the underlying script:
https://github.com/WarnerMedia/artemis/blob/main/backend/engine/plugins/github_repo_health/cli/README.md#rules