Store secret types in the database instead of as a hardcoded list (#30)
The ghas_secrets plugin introduced a large number of new secret types
with the high likelihood of new types being added in the future. This
created a problem where the previous hardcoded list of secret types for
API filtering validation was no longer valid.

Instead of continuing to maintain a static list of secret types this
change moves the list into the database. When secret detection plugins
finish any new types are added to the database. The API uses the list
from the database for filter validation. The database table is
pre-populated with the current static list of types.