Fix bug in report generation with config results
This fixes an issue where reports would fail to generate if the scan had
configuration results and configuration allowlist items. Because the
"id" key was removed in https://github.com/WarnerMedia/artemis/pull/31
it was not present to check against the allowlist items, causing an
exception. This fix puts the "id" key back for the allowlist check.