Make plugin event processing environment-configurable (#40)
While the engine can process plugin events for secrets, inventory, and
configuration plugins not every deployment is going to have need for
them. Rather than generating the events and letting them get dropped at
the event dispatch lambda, which might generate log messages depending
on the event type, make it so each type can be turned on or off. This
way only the events that are used will be generated.