New Feature: Email Aliases for Authentication (#50)
This feature handles a situation where users' email addresses have changed in a graceful manner. Rather than go into the database and manually change email addresses as they are updated, risking downtime for some users, a variable can be defined with a mapping of the old email format to the new email format. When a user logs in with their new email, the existing user with the old email will be found and automatically updated with the new email.