Migrate away from obsolete template provider.
The "template_file" data source uses the hashicorp/template provider,
which is officially deprecated and archived, with the functionality
moved into the templatefile function in Terraform 0.12.

This is a problem because the final release of the template provider,
2.2.0, does not include a darwin_arm64 build, so folks on that platform
will get the following error:

    template v2.2.0 does not have a package available for your current platform, darwin_arm64

The official migration path is to use the templatefile function instead.

This change only removes the dependency on the hashicorp/template
provider, there are no functional changes.

See also:
https://registry.terraform.io/providers/hashicorp/template/latest/docs#deprecation