Improve unit test coverage and isolation (#130)
* Include ProfileNotFound in AWS credential check.

If an AWS profile is set via env vars but is not found in the
credentials source, then the error will be ProfileNotFound rather than
NoCredentialError.

* Add unit tests for get_all_metadata.

* Add unit test for load_schemes().

* Mock AWSConnect() to prevent real AWS connection.

Even though we stub out the lambda_execute, the constructor still
attempts to load the AWS credentials, which fails if no profile is
available.

* Patch out the AuditLogger in tests.

The AuditLogger sets up an AWS connection based on the current
environment vars, which fails if no profile is available.

* Add tests for imported modules.