Fix coverage report.
This adjusts the coverage tracking to match the pytest search paths,
which fixes the lambdas not being included as well as the error about
the "api" module never being included.

Additionally, we filter out paths which we don't track toward coverage
metrics, namely the tests themmselves and some utility scripts.