Backend test consistency fixes and cleanups (#131)
* Remove venv target.

This target was only used a dependency of the test targets and was
responsible for installing/upgrading pipenv.

Pipenv is already expected to be installed, so this step just added time
and noise when running tests, and we want to make running tests faster
and easier.

* Add step to install Python deps for tests.

* Add missing .PHONY targets for test targets.

* Remove "pipenv check" step from test targets.

We regularly run Artemis on itself, so this step is unnecessary and adds
noise to running the tests.

"pipenv check" can still be run manually.

* Include json_report in test PYTHONPATH.

The json_report lambda was being included directly in the unit-test and
integration-test targets, but that meant that the end2end-test and test
targets would fail.

This adds it to COMMON_PYTHONPATH so that it's consistently included in
all test targets.

* Don't load the .env file when running tests.

The values in the .env file should not affect tests. This improves
consistency between tests run locally versus run via CI.

* Fix local Trivy installation.

The get-trivy target was hard-coded to install Trivy into
/usr/local/bin. This changes it to install into a relative directory,
similar to how we install other test tools such as shellcheck.

* Prefer our downloaded tools in tests.

Before, if a locally-installed tool was in the PATH, then that instance
would be used instead of the specific versions we're downloading. This
changes the search priority to prefer our tools.

* Refactor test steps into common define.

* Fix coverage report.

This adjusts the coverage tracking to match the pytest search paths,
which fixes the lambdas not being included as well as the error about
the "api" module never being included.

Additionally, we filter out paths which we don't track toward coverage
metrics, namely the tests themmselves and some utility scripts.