Orchestrator integration test fixes (#138)
* Fix test file location and expected service list.

This test was attempting to pull the services.json from a real S3 bucket
and was likely marked as an integtest because of it. However,
SERVICES_LOC was defined as well which would suggest it was supposed to
be a regular unit test instead.

For now, we are changing it to a regular unit test.  If we want to test
S3 functionality in the future (which we do), then we can mock it.

* Add git.example.com service.

This service is referenced in the integration tests.

* Pass ARTEMIS_API to integration tests.

This fixes an issue with test_org_queue.

* Remove non-functional no-pylint flag.

This fixes the "pytest: error: unrecognized arguments: --no-pylint"
error.