Upgrade to pytest v8 and move PYTHONPATH config (#142)
* Upgrade to Pytest v8.

* Add psycopg2-binary to dev deps.

This allows the unit tests to run from a clean venv without explicitly
installing artemisdb (and also specifies the minimum version).

* Set PYTHONPATH in tests via pytest.ini.

This is working towards moving everything needed to run tests into
pytest configuration to make automation and IDE integration easier.

We also remove the step to install artemislib and artemisdb since it is
now included in the PYTHONPATH.