Update Python dependencies (2024-05-20) (#146)
* Set conservative version ranges.

This is to allow us to more carefully upgrade dependencies to the newest
major versions a piece at a time.

Utilities which are used in plugins are locked at specific versions
since those will need extra care.

black and moto are locked at specific versions since they will be
upgraded to newer major versions in a later update.

Also removed unused pytest-black and sorted the lists alphabetically.

* Update deps.

autopep8 and flake8 upgraded to latest versions due to dependency
conflict.

* Document why cfn-lint and checkov are locked.

* Remove pytest-cache and pytest-logging.

These plugins are now part of pytest itself.

* Upgrade pytest plugins.

* Update cryptography and mock.

* Lock packaging to match Makefile.

* Harmonize deps with backend.

* Harmonize version ranges with Pipfile.

* Move psycopg2-binary to main deps.

This dep is a direct requirement of artemisdb.

* Add python-jose dep since it's used by authorizer.

* Harmonize dependency versions with Pipfile.

* Update deps.