UI test enhancements (#161)
* Skip "export fails" tests.

These tests are failing unpredictably.  Tested by running the tests
repeatedly on both CI and my dev boxes.  This appears to have been an
issue for a while now, unrelated to the recent dependency updates.

* Generate HTML test report.

The UI tests generate an exceptional amount of output. Updating
dependencies often cause a large number of errors which are difficult to
review via the logs.

This HTML output makes it easier to view the test results and can be
captured as a CI artifact.

* Generate HTML report on "test" / "test-precommit".

* Include console log if available.

* Save HTML test report as an artifact.

* Use default retention.

The size of the HTML report is tiny, so storage size isn't a concern.

Additionally, all of the information in the report is also in the logs.

* Fix whitespace.