Fetch Detekt during build of the Java images (#174)
* Fetch Detekt during build of the Java images.

This is consistent with how we fetch dependent tools; we do not rely on
the artifacts from the unit tests and vice-versa.

The long-term plan is for the unit tests to not be dependent on the
Makefile at all.

* Verify detekt.jar SHA-256 hash.

* Restore embedded version in Detekt path.

The Java images no longer depend on the unit test tool download.