Preparation for future tool dependency updates (#175)
* Ignore order of findings and message details.

This prepares us for more regular updates of Checkov in the future (in
particular 3.2) by allowing us to assert on specific findings, without
caring if the order changes or the text of the messsage is adjusted.

For example, newer versions of Checkov use different URLs in their
messages.

* Add trufflehog as a dev dep.

This is to detect potential dependency conflicts in Dockerfile.python3.

* Clean up lint warnings.