Enable grouped Dependabot PRs for GitHub Actions.
This enables generating PRs by Dependabot for updates to the GitHub
Actions workflows.

To keep the number of PRs low, we group all updates from common
low-risk official sources into a single PR.  This makes it easier to
review and merge during our biweekly dependency update cycle.

We are only focusing on GitHub Actions for now since we pin to version
hashes in our workflows, which is difficult to do correctly manually.