Download tools in a common stage.
OWASP FindSecBugs and Detekt are included in all Java images, so we now
download them into a common build stage which can be cached and shared
between dist targets.

Dependency Check is not included since it is currently only enabled for
the Java 8 image, and we don't know if the "updateonly" step is portable
between Java versions.