Build all Java images in parallel.
The definition for building the Java images has been moved to a
Bakefile.  This simplifies the build definition by allowing us to use
HCL and functions to reduce the repetition.

For now, the ability to build individual Java images via the Makefile
has been removed, but we can add it back in if necessary by breaking out
the "deploy-java" target into a generic function.

Building the "java" group in the Bakefile builds all of the images in
parallel, including sharing the common tool download stage.