Streamline Java image build (#207)
* Install Alpine deps in one go.

* Set up Python without Pip cache.

* Streamline FindSecBugs installation and clean up.

* Silence Hadolint warning about pipefail.

We are using Alpine's /bin/sh so ignoring the warning is the recommended
course of action.

* Remove Dependency Check install temp files.

* Specify whitespace for HCL.

* Download tools in a common stage.

OWASP FindSecBugs and Detekt are included in all Java images, so we now
download them into a common build stage which can be cached and shared
between dist targets.

Dependency Check is not included since it is currently only enabled for
the Java 8 image, and we don't know if the "updateonly" step is portable
between Java versions.

* Build all Java images in parallel.

The definition for building the Java images has been moved to a
Bakefile.  This simplifies the build definition by allowing us to use
HCL and functions to reduce the repetition.

For now, the ability to build individual Java images via the Makefile
has been removed, but we can add it back in if necessary by breaking out
the "deploy-java" target into a generic function.

Building the "java" group in the Bakefile builds all of the images in
parallel, including sharing the common tool download stage.

* Move docker-bake.hcl up to reflect the context.

* Support single-image Java docker builds.

This restores support for the single-image-at-a-time targets for the
Java images.  We use an implicit target in the Makefile since the
bakefile already defines what targets are available.

This allows us to support multiple types of CI workflows, and gradual
migrations between them:

* Workflows that build all Docker images at once ("docker" target).
* Workflows that build all images one at a time ("dist/docker/*"
  targets), possibly in parallel jobs.
* Workflows that use a hybrid of the above approaches ("docker-java" for
  Java images, "dist/docker/*" for the others).