Include boto3-stubs in artemislib lambda layer.
Even though boto3-stubs only includes typing info, it is still needed to
be included in the main (non-dev) packages in order to resolve imports
at runtime.

This also brings the Makefile up-to-date with the artemislib
dependencies.

boto3 itself is also now included in the layer so that we use the
version we specify rather than whatever version is bundled with the
lambda runtime.