Improve handling of API errors. (#256)
This improves our handling of cases where the REST API calls return an
error, particularly non-JSON responses.

Before, these conditions would cause the script to exit immediately with
a jq error message. Now we provide a proper diagnostic message to aid in
debugging.

Now when printing the response body, if the body text cannot be
pretty-printed as JSON, we fall back to printing the raw body. We also
add default values for when extracting a JSON value fails.