Fix CI build by restricting setuptools to pre-v74. (#260)
setuptools v74 is causing a craash when pip installs dependencies with
the error:

    TypeError: canonicalize_version() got an unexpected keyword argument
    'strip_trailing_zero'

The workaround is to constrain setuptools to v73 or earlier. We must
also pin virtulenv since 20.26.4 is hard-coded to install setuptools v74
into new virtual environments.

We can likely remove this restriction when we can upgrade to packaging
v24.