Adjust PluginSettings types and add unit tests (#282)
* Adjust PluginSettings types and add unit tests.

* Improve validation of PluginSettings.

PluginSettings is now validated via Pydantic, which enforces stricter
type checks at runtime.

* Move PluginSettings.name to top.

This is the only field which is required (everything else has default
values). Also added a docstring to point it out.

* Add test for JSON type checking.

* Fix ValueError if "lib" dir doesn't exist.

* Include build_images in PluginSettings.

EngineProcessor no longer needs to parse the settings.json itself; it
can use get_plugin_settings() and take advantage of the validation
checks.

* Use test data for testing docker_images_required.

This removes the unit test dependency on the actual plugin settings,
which means we can also remove the xfail.

The names of the test functions have been also fixed ("image" ->
"images") to match the function under test.

* Simplify is_plugin_disabled.

This function is now only used by the PluginSettings validator, so we
don't need to create a temporary dict anybody.

* Move is_plugin_disabled into PluginSettings.

The PluginSettings validator is the only remaining user of this
function.