Build a self-contained core plugin bundle.
This is an alternative method of running plugins within a container.

Instead of relying on the system Python and dependencies in the target
container, this bundles the plugins, dependencies, and a "portable"
Python distribution into a single self-contained executable.

This means that the target container no longer needs a system Python
installed and we don't need to make sure the dependencies are installed
in the container (which can get out of sync with the engine
dependencies).

The only new requirement of the target container is that it must be a
glibc-based distribution since the bundle contains a glibc-based Python
and shared libraries.