New plugin runner: boxed (#299)
* Build a self-contained core plugin bundle.

This is an alternative method of running plugins within a container.

Instead of relying on the system Python and dependencies in the target
container, this bundles the plugins, dependencies, and a "portable"
Python distribution into a single self-contained executable.

This means that the target container no longer needs a system Python
installed and we don't need to make sure the dependencies are installed
in the container (which can get out of sync with the engine
dependencies).

The only new requirement of the target container is that it must be a
glibc-based distribution since the bundle contains a glibc-based Python
and shared libraries.

* Support plugins using the new "corebox" runner.

* Switch golang plugins to use new "corebox" runner.

The golang Dockerfile no longer requires any extra dependencies to be
installed, but it does need to be a glibc-based distro (in this case,
Debian Buster).

* Fix path priority.

Make sure our Entry and Gosec binaries take precendence over any other
system binaries that might be in the PATH.  This isn't an issue with our
current base image, but this prevents future potential issues.

Also fixes the Docker build warning about the old ENV syntax.

* Change "corebox" to "boxed".

* Add documentation.

* Add "boxed" plugin runner.

* Document our minimal base image reqs.

* Refactor plugin name validation.

Moved the plugin name validation out from the runner into the a util
function so we can better document and unit test it (and re-use it in
some other places).

* Use validate_plugin_name in get_plugin_list().

This fixes the issue that the "__pycache__" directory is being registered
as a plugin.

* Fix circular import.

Importing engine.plugins.lib.utils is causing a circular import issue at
runtime when the engine starts up, so for now we use a simplified
version of the plugin name check.

* Document the "boxed" runner.

* Add note about not using validate_plugin_name().