Fix return value and test data for parse_stderr().
parse_stderr() should always return a list[str] and not a bare string.

Additionally, the test data in the unit test had extraneous
word-wrapping (perhaps an inadvertant reformat?) and so did not reflect
what the stderr output would look like.  The new unit test data is a
simplified amalgamation based on examples of logs posted online.