Handle error messages other than ENOENT.
The previous processing of stderr was focused on a single case which
was the Gemfile.lock not being found, which meant that other errors were
being ignored or mangled.

This reworks the stderr processing to take a "lighter" touch by only
trimming Ruby stack trace lines beyond the first line, and even then not
doing any further processing on that line.  This avoids potentially
missing any error messages with the side-effect being that actual stack
traces are more verbose to the user.

The good news is that triggering a stack trace is much rarer with modern
versions of bundler-audit, since most common error messages are properly
formatted as a simple error message without a stack trace.