Clean up bundler-audit plugin (#323)
* Fix return value and test data for parse_stderr().

parse_stderr() should always return a list[str] and not a bare string.

Additionally, the test data in the unit test had extraneous
word-wrapping (perhaps an inadvertant reformat?) and so did not reflect
what the stderr output would look like.  The new unit test data is a
simplified amalgamation based on examples of logs posted online.

* Handle error messages other than ENOENT.

The previous processing of stderr was focused on a single case which
was the Gemfile.lock not being found, which meant that other errors were
being ignored or mangled.

This reworks the stderr processing to take a "lighter" touch by only
trimming Ruby stack trace lines beyond the first line, and even then not
doing any further processing on that line.  This avoids potentially
missing any error messages with the side-effect being that actual stack
traces are more verbose to the user.

The good news is that triggering a stack trace is much rarer with modern
versions of bundler-audit, since most common error messages are properly
formatted as a simple error message without a stack trace.