Overhaul handling of tool errors.
Parsing of stdout and stderr from running bundler-audit is now more
consistent, with better detection of error conditions.

A previous fix adjusted the stderr parsing so that it no longer
ignores/mangles error messages, but this exposed the issue that git
(which is called by bundler-audit) writes its progress messages to
stderr, which was being captured as errors.

We now check the exit code of the tool as well as if any findings are
present in order to determine if the errors should be reported.

A number of test cases were added based on actual test runs of
bundler-audit so that the tests better reflect reality.