Overhaul handling of errors from bundler-audit (#338)
* Overhaul handling of tool errors.

Parsing of stdout and stderr from running bundler-audit is now more
consistent, with better detection of error conditions.

A previous fix adjusted the stderr parsing so that it no longer
ignores/mangles error messages, but this exposed the issue that git
(which is called by bundler-audit) writes its progress messages to
stderr, which was being captured as errors.

We now check the exit code of the tool as well as if any findings are
present in order to determine if the errors should be reported.

A number of test cases were added based on actual test runs of
bundler-audit so that the tests better reflect reality.

* Omit empty findings from results.

If convert_dict could not parse the block then it would return an empty
dict which would end up in the list of findings, which would then need
to be accounted for elsewhere.

* Only report the last line from stderr.

* Handle special "failed to download" error.

* Add note backtrace format.

Specifically, we do not need to deal with Ruby 2.5 backtrace format,
which reverses the order of lines and adds numbering. This format is
only activated when the output is a TTY.

* Remove duplicate log entry.