Generate and deploy API documentation with UI (#357)
* Add Redocly-CLI.

* Move spec.yaml into ui tree.

Generation of the API documentation has been moved to the UI build
process, so the spec file needs to be contained within the UI source
tree so the CI build can access it.

We add a symlink in the old location for convenience.

* Serve static API doc files.

* Add redocly-lint target.

* Clean generated API docs.

* Remove API docs generation.

This has been moved to the UI build process.

* Deploy api/index.html as api/.