Move temp vol mount and pass mount point.
Moved the temporary named volume mount point from /work/tmp to
/tmp/work. If the container mounts /work as read-only, then Docker can't
mkdir /work/tmp.  /tmp is a safer choice since it should always be
writable.

We also now pass the mount point in the temp_vol_name engine var so that
if we change it again later, plugins will be able to adapt.