Provide a temporary "scratch" volume to plugins (#363)
* Rename "docker" to "oci".

Side-step name conflict with Docker SDK for Python.

* Add Docker SDK.

* Fix crash on invalid JSON in secret.

* Provide a temp named volume to each plugin.

The temporary volume allows containers-within-containers to share a
writable "scratch space" which is deleted after the plugin completes.
Otherwise, if a container launches another container, it would need to
know what is safe to bind mount from the host (since bind mounts are
always mounted from the host system).

* Mount target dir with same path in container.

This more closely matches the way the deployed engine container
mounts the "/work" directory from the host.

This is important for plugin containers that want to mount the source
directory into a sub-container.

* Provide a temporary volume to the plugin.

* Move temp vol mount and pass mount point.

Moved the temporary named volume mount point from /work/tmp to
/tmp/work. If the container mounts /work as read-only, then Docker can't
mkdir /work/tmp.  /tmp is a safer choice since it should always be
writable.

We also now pass the mount point in the temp_vol_name engine var so that
if we change it again later, plugins will be able to adapt.

* Use plugin ID instead of name in vol name.

* Mount target dir in engine and inherit.

We now mount the target dir in the engine and inherit the volume in the
plugin container. This is much closer to the deployed engine and denies
the assumption that the host path matches the mount path.

* Revert target path examples.

The mount dir is no longer matched to the host dir.

* Pass the engine ID.