Write custom config to temp volume.
The original implementation for Checkov custom configs wrote the config
files to the plugin source directory. This would get persisted between
scans that get assigned to the same engine instance.

We now use the temporary volume to provide a consistent location for
passing configs to the Checkov container.

Additionally, this simplifies the logic around the custom severities
file by giving it a consistent location in the temporary volume.