Colorize stderr in plugin output.
This replaces the shell-based plugin wrapper with a statically-compiled
Go wrapper which monitors stdout and stderr.
The output lines are serialized (i.e., if the plugin writes to stdout
and stderr at the same time, the lines will be written separately) and
colorized.

This sets us up to implement more useful debug tooling such as result
linting, logging, timeouts, etc.