Add Plugin Runner Toolbox (#371)
* Support containers without a shell.

This adds a "toolbox" container that, for now, just provides a BusyBox
toolset so that the plugin startup script can run. It also provides a
debug shell if the container doesn't ship with one.

* Colorize stderr in plugin output.

This replaces the shell-based plugin wrapper with a statically-compiled
Go wrapper which monitors stdout and stderr.
The output lines are serialized (i.e., if the plugin writes to stdout
and stderr at the same time, the lines will be written separately) and
colorized.

This sets us up to implement more useful debug tooling such as result
linting, logging, timeouts, etc.

* Add basic linting of plugin results.

* Build the toolkit separately in quit mode.

The build process was generating quite a bit of output on every run.

* Fix lint check for "truncated" field value.

* Be consistent on "toolbox" (not "toolkit").

* Add comment about run-plugin.

* Fix truncated comment.