Handle large output lines from the plugin.
Previously, output lines were assumed to fit within the default Scanner
mmximum buffer size of 64 KB.  This has quickly been determined to be
overly optimistic!

Refactored the line scanning into scanPipe(), with a new maximum line
buffer size of 128 MB.  If a line exceeds this threshold, then we
display an error and forcefully close the pipe in order to avoid a
deadlock.