Re-enable default caching for docker build.
Removes the "--no-cache" and "--force-rm" options from always being
set. These can be set using the DOCKER_BUILD_EXTRA_ARGS var on an
as-needed basis.

We want to make use of both layer caches and cache mounts when building
our container images to improve the local dev cycle.  There may be
additional tweaks needed to .dockerignore to avoid unnecessary rebuilds,
but this is a first start.

This will not affect CI deploys since we assume that each CI run will
start with an empty cache anyway.