Handle null URL in secrets results. (#383)
Fixes the "Hide Finding" dialog for secrets findings.

Typically the "url" field of a secrets finding is left blank by being
set to an empty string. However, sometimes the service returns null
instead.

This is not currently specified in the API spec, so for now we will
handle both cases.