Re-enable default caching for docker build (#380)
* Re-enable default caching for docker build.

Removes the "--no-cache" and "--force-rm" options from always being
set. These can be set using the DOCKER_BUILD_EXTRA_ARGS var on an
as-needed basis.

We want to make use of both layer caches and cache mounts when building
our container images to improve the local dev cycle.  There may be
additional tweaks needed to .dockerignore to avoid unnecessary rebuilds,
but this is a first start.

This will not affect CI deploys since we assume that each CI run will
start with an empty cache anyway.

* Enable cache mounts for apt, pip, and pipenv.

* Ignore .temp directory.

In the past this was included since some Dockerfiles would copy the
cached tools from the unit tests into the container image.

We no longer do this (Dockerfiles always download the tools themselves),
so this is no longer necessary.

* Ignore coverage report and image directories.

* Lock cache mounts for Apt and Pipenv.

Apt definitely needs sharing=locked, but Pipenv is unknown so we set it
to be safe.