Pre-extract the boxed plugin bundle (#386)
* Pre-extract the boxed plugin bundle.

Addresses two issues:

* Improves boxed plugin startup time.
* Lessens chance of free disk space errors, such as "Consider setting
  TMPDIR to a directory with more free space".

This is part of the longer-term effort to support distributions based on
Musl libc (such as Alpine) by moving the glibc-based bundle into it's
own directory. The bundle build process is now its own Dockerfile so we
can re-use it (via docker-bake.hcl) to build both Glibc and Musl
variants.

* Fix indentation.

* Rename ".boxed" to "_boxed".

Fixes engine startup issue where it was trying to auto-register the
directory as a plugin.