Ignore Powertools warning about no metrics.
This warning is generated by AWS Lambda Powertools when there are no
metrics generated during the handler. In the unit tests, this is normal
since we have test cases where the handler effectively does nothing.

Note that we already filter this warning in some of the lambdas -- this
config is needed to also filter the warnings at the Pytest level.