Disable word-wrap option in raw view.
prism-react-renderer does not support word-wrapping itself, and even
with the old viewer the word-wrapping functionality wasn't great, so
for now we'll disable this feature.

We may revisit this if this is a requested feature, but it's generally a
better idea to copy and paste the JSON into a proper JSON editor anyway.