Switch raw JSON view to prism-react-renderer (#477)
* Switch raw view to prism-react-renderer.

react-syntax-highlighter holds back updating prismjs to address a
potential security vulnerabliity.

* Fix light/dark classification of "github" theme.

* Add missing className to syntax highlight.

* Toggle line numbers via colgroup.

* Reformat.

* Disable word-wrap option in raw view.

prism-react-renderer does not support word-wrapping itself, and even
with the old viewer the word-wrapping functionality wasn't great, so
for now we'll disable this feature.

We may revisit this if this is a requested feature, but it's generally a
better idea to copy and paste the JSON into a proper JSON editor anyway.

* Scroll the raw viewport separately.

This matches the previous implemention but we may want to revisit this
since it means the scroll bar is at the bottom of the container rather
than the browser window.

* Simplify raw tab styles and move into classes.

* Adjust line number color.

We don't have a separate theme color for the line numbers, so we adjust
the opacity of the default theme foreground color.

* Update unit tests for prism-react-renderer.

* Updated the referenced themes.
* Commented-out "wrap long lines" checks since that is currently
  disabled.

* Remove @types/react-syntax-highlighter.

* Bump to 1.16.0.