Eject eslint-config-react-app.
Removes the dependency on eslint-config-react-app (which is no longer
maintained) by adapting the configuration into our own .eslintrc.json.

This unblocks us from upgrading many of the plugins as well as migrating
to ESLint v9. Additionally, we can now more easily experiment with
adjusting the rules to fit our style.