Migrate to ESLint v9 (#492)
* Eject eslint-config-react-app.

Removes the dependency on eslint-config-react-app (which is no longer
maintained) by adapting the configuration into our own .eslintrc.json.

This unblocks us from upgrading many of the plugins as well as migrating
to ESLint v9. Additionally, we can now more easily experiment with
adjusting the rules to fit our style.

* Clean up typescript-eslint/no-unused-vars.

This rule was introduced when we upgraded typescript-eslint/no-unused-vars.

* Update translations.

* Migrate to ESLint v9.

* Clean up unused ESLint directives.

These unused directives are now warnings in ESLint v9.

* Update translations.