Upgrade Go in Docker images to 1.24.5 (#501)
* Migrate download stage to Alpine 3.22.

* Switch from Debian to Alpine.

This reduces the image size from 1.2 GB to 475 MB.

* Upgrade to gosec 2.22.7.

This fixes compatiblity with Go 1.24.

* Upgrade to Go 1.24.

* Remove glibc requirement.

* Install Go 1.24.5 from golang image.

The Go distribution is self-contained, so we can upgrade Go separately
from what is available in the Docker DinD distro.

* Improve type validation.

Merges changes from the gosec-plugin-typing branch used while debugging
Go 1.23 issues.

* Temporarily use Python 3.13 in boxed-musl.

This works around a runtime error with Python 3.12 on Musl:
ModuleNotFoundError: No module named 'pydantic_core._pydantic_core'

Other imports are affected; Pydantic is just the most common we've
seen.

This fix has not yet been backported to a Python 3.12 release, so we
will install 3.13 for now.

* Fix missing required output fields.

Also skip the stringify step and serialize the object directly to
stdout.