Overhaul Docker ImageBuilder types and clean up build cache (#512)
* Clean up docstrings and typehints.

Define TypedDicts for the dicts which are passed around to document the
fields and ensure consistent usage.

The return value of remove_docker_image() was simplified. The return
value isn't actually checked or used anywhere, though.

* Move comment to correct location.

Also fixed typo.

* Refactor remove_docker_image.

The preferred way to remove an image is via BuiltImage.remove(). This is
so that we can include additional metadata in the future.

* Prune images *after* removing images.

* Add typehints for BuildResult.

* Upgrade BuiltImage and BuildResult to Pydantic.

This enables full JSON serialization and deserialization with
typechecking, so we can in the future move them to artemislib so plugins
can make use of them.

* Rename BuildResult to ScanImages.

This matches the arg name used when passing it around.

* Add docstrings for attributes.

* Include test_build_local_image in standard tests.

We already require Docker to run unit tests, so we can include this in
the standard unit tests and not in integration tests.

To keep the test fast and consistent, we instead build a simple
Dockerfile that does not rely on any external dependencies which might
fail to fetch.

* Build images in a temporary builder.

When the temporary builder is removed, then all associated resources
(i.e. the build cache) are released, freeing up disk space.

* Prefix builder name with "artemis".

Builder names must start with a letter. The engine ID on its own is just
a series of hex digits.

* Fix built image not being loaded.

* Add additional unit tests for build_local_image.