Upgrade pytest-subtests to 0.14.2 (#524)
* Upgrade pytest-subtests to 0.14.2.

* Fix subtests with xdist in pytest-subtest 0.14.2.

pytest-subtest now reports each subtest individually. The title of each
test is generated by serializing the subtest values, so any values that
can't be trivially serialized (such as datetime.datetime) can't be
reported when using xdist.

This makes sure that all subtest cases are trivially serializable.

* Add warning that the req arg may be modified.

Callers which expect the req to be unchanged must pass a deep-copy.

* Fix comment typo.